+++ extras/AFC.py
@@ -10,7 +10,7 @@ import traceback
 from configfile import error
 from typing import Any
 
-from typing import Dict, TYPE_CHECKING
+from typing import Dict, TYPE_CHECKING, Union
 
 if TYPE_CHECKING:
     from configfile import ConfigWrapper
@@ -21,10 +21,11 @@ if TYPE_CHECKING:
     from extras.AFC_hub import afc_hub
     from extras.AFC_spool import AFCSpool
     from extras.AFC_error import afcError
+    from extras.AFC_stepper import AFCExtruderStepper
 
 ERROR_STR = "Error trying to import {import_lib}, please rerun install-afc.sh script in your AFC-Klipper-Add-On directory then restart klipper\n\n{trace}"
 
-try: from extras.AFC_lane import AFCLaneState, SpeedMode, AssistActive
+try: from extras.AFC_lane import AFCLaneState, SpeedMode, AssistActive, AFCHomingPoints
 except: raise error(ERROR_STR.format(import_lib="AFC_logger", trace=traceback.format_exc()))
 
 try: from extras.AFC_logger import AFC_logger
@@ -98,7 +99,7 @@ class afc:
         # Objects for everything configured for AFC
         self.units      = {}
         self.tools: Dict[str, AFCExtruder] = {}
-        self.lanes: Dict[str, AFCLane]     = {}
+        self.lanes: Dict[str, Union[AFCLane, AFCExtruderStepper]] = {}
         self.hubs       = {}
         self.buffers    = {}
         self.tool_cmds  = {}
@@ -202,6 +203,9 @@ class afc:
         self.tool_homing_distance   = config.getfloat("tool_homing_distance", 200)  # Distance over which toolhead homing is to be attempted.
         self.max_move_dis           = config.getfloat("max_move_dis", 999999)       # Maximum distance to move filament. AFC breaks filament moves over this number into multiple moves. Useful to lower this number if running into timer too close errors when doing long filament moves.
         self.n20_break_delay_time   = config.getfloat("n20_break_delay_time", 0.200)# Time to wait between breaking n20 motors(nSleep/FWD/RWD all 1) and then releasing the break to allow coasting.
+        self.home_to_hub            = config.getboolean("home_to_hub", True)        # Global setting to auto-home to hub during moves
+        self.home_to_tool           = config.getboolean("home_to_tool", True)       # Global setting to auto-home to tool during moves
+        self.homing_enabled         = config.getboolean("homing_enabled", True)
 
         self.tool_max_unload_attempts= config.getint('tool_max_unload_attempts', 4) # Max number of attempts to unload filament from toolhead when using buffer as ramming sensor
         self.tool_max_load_checks   = config.getint('tool_max_load_checks', 4)      # Max number of attempts to check to make sure filament is loaded into toolhead extruder when using buffer as ramming sensor
@@ -428,7 +432,7 @@ class afc:
             pin_loc_xy = tool_cut_obj.variables.get('pin_loc_xy', None)
             if pin_loc_xy and pin_loc_xy == (-99,-99):
                 error_str += 'tool_cut is set to True and variable_pin_loc_xy has not been updated.\n'
-                error_str += 'Please update variable_pin_loc_xy in AFC\AFC_Macro_Vars.cfg file.\n\n'
+                error_str += 'Please update variable_pin_loc_xy in AFC\\AFC_Macro_Vars.cfg file.\n\n'
 
         park_obj = self.printer.lookup_object('gcode_macro _AFC_PARK_VARS', None)
         if (self.park
@@ -437,7 +441,7 @@ class afc:
             park_loc_xy = park_obj.variables.get('park_loc_xy', None)
             if park_loc_xy and park_loc_xy == (-99,-99):
                 error_str += 'park is set to True and variable_park_loc_xy has not been updated.\n'
-                error_str += 'Please update variable_park_loc_xy in AFC\AFC_Macro_Vars.cfg file.\n\n'
+                error_str += 'Please update variable_park_loc_xy in AFC\\AFC_Macro_Vars.cfg file.\n\n'
 
         poop_obj = self.printer.lookup_object('gcode_macro _AFC_POOP_VARS', None)
         if (self.poop
@@ -446,7 +450,7 @@ class afc:
             purge_loc_xy = poop_obj.variables.get('purge_loc_xy', None)
             if purge_loc_xy and purge_loc_xy == (-99,-99):
                 error_str += 'poop is set to True and variable_purge_loc_xy has not been updated.\n'
-                error_str += 'Please update variable_purge_loc_xy in AFC\AFC_Macro_Vars.cfg file.\n\n'
+                error_str += 'Please update variable_purge_loc_xy in AFC\\AFC_Macro_Vars.cfg file.\n\n'
 
         kick_obj = self.printer.lookup_object('gcode_macro _AFC_KICK_VARS', None)
         if (self.kick
@@ -455,7 +459,7 @@ class afc:
             kick_start_loc = kick_obj.variables.get('kick_start_loc', None)
             if kick_start_loc and kick_start_loc == (-99,-99,5):
                 error_str += 'kick is set to True and variable_kick_start_loc has not been updated.\n'
-                error_str += 'Please update variable_kick_start_loc in AFC\AFC_Macro_Vars.cfg file.\n\n'
+                error_str += 'Please update variable_kick_start_loc in AFC\\AFC_Macro_Vars.cfg file.\n\n'
 
         wipe_obj = self.printer.lookup_object('gcode_macro _AFC_BRUSH_VARS', None)
         if (self.wipe
@@ -464,7 +468,7 @@ class afc:
             brush_loc = wipe_obj.variables.get('brush_loc', None)
             if brush_loc and brush_loc == (-99,-99,-1):
                 error_str += 'wipe is set to True and variable_brush_loc has not been updated.\n'
-                error_str += 'Please update variable_brush_loc in AFC\AFC_Macro_Vars.cfg file.\n\n'
+                error_str += 'Please update variable_brush_loc in AFC\\AFC_Macro_Vars.cfg file.\n\n'
 
         return error_str
 
@@ -1038,12 +1042,23 @@ class afc:
         cur_lane.status = AFCLaneState.HUB_LOADING
         if not cur_lane.load_state:
             while not cur_lane.load_state:
-                cur_lane.move_advanced( cur_hub.move_dis, SpeedMode.SHORT)
+                # TODO: add timout routine here
+                cur_lane.move_to(cur_hub.move_dis, SpeedMode.SHORT,
+                                 endstop=AFCHomingPoints.LOAD,
+                                 assist_active=AssistActive.DYNAMIC,
+                                 use_homing=self.homing_enabled)
         if not cur_lane.loaded_to_hub:
-            cur_lane.move_advanced(cur_lane.dist_hub, SpeedMode.HUB, assist_active = AssistActive.DYNAMIC)
+            dist_to_hub = cur_lane.dist_hub
+            if self.homing_enabled:
+                dist_to_hub = cur_lane.dist_hub + cur_hub.move_dis
+            cur_lane.move_to(dist_to_hub, SpeedMode.HUB, assist_active=AssistActive.DYNAMIC,
+                             endstop=AFCHomingPoints.HUB, use_homing=self.homing_enabled)
         while not cur_hub.state:
-            cur_lane.move_advanced(cur_hub.move_dis, SpeedMode.SHORT)
+            # TODO: add timout routine here
+            cur_lane.move_to(cur_hub.move_dis, SpeedMode.SHORT, assist_active=AssistActive.NO,
+                             endstop=AFCHomingPoints.HUB, use_homing=self.homing_enabled)
         while cur_hub.state:
+            # TODO: add timout routine here
             cur_lane.move_advanced(cur_hub.move_dis * -1, SpeedMode.SHORT)
         cur_lane.status = AFCLaneState.NONE
         cur_lane.do_enable(False)
@@ -1082,6 +1097,8 @@ class afc:
         self.LANE_UNLOAD( cur_lane )
 
     def LANE_UNLOAD(self, cur_lane: AFCLane):
+        # TODO: update this to unload from toolhead and move all the way back to load
+        # when homing is enabled
         cur_hub = cur_lane.hub_obj
 
         self.current_state = State.EJECTING_LANE
@@ -1096,11 +1113,16 @@ class afc:
             cur_lane.status = AFCLaneState.EJECTING
             self.save_vars()
             if cur_lane.loaded_to_hub:
-                cur_lane.move_advanced(cur_lane.dist_hub * -1, SpeedMode.HUB, assist_active = AssistActive.DYNAMIC)
+                cur_lane.move_to( cur_lane.dist_hub * -1, SpeedMode.HUB,
+                                  endstop=AFCHomingPoints.LOAD,
+                                  assist_active=AssistActive.DYNAMIC,
+                                  use_homing=self.homing_enabled)
             cur_lane.loaded_to_hub = False
             while cur_lane.load_state:
-                cur_lane.move_advanced(cur_hub.move_dis * -1, SpeedMode.SHORT, assist_active = AssistActive.YES)
-            cur_lane.move_advanced(cur_hub.move_dis * -5, SpeedMode.SHORT)
+                # TODO: add timout routine here
+                cur_lane.move_advanced(cur_hub.move_dis * -1, SpeedMode.SHORT,
+                                       assist_active = AssistActive.YES)
+            cur_lane.move_advanced(cur_lane.extruder_clear_dis * -1, SpeedMode.SHORT)
             cur_lane.do_enable(False)
             cur_lane.status = AFCLaneState.NONE
             cur_lane.unit_obj.return_to_home()
@@ -1318,18 +1340,36 @@ class afc:
 
             # Move filament to the hub if it's not already loaded there.
             if not cur_lane.loaded_to_hub or cur_lane.is_direct_hub():
-                cur_lane.move_advanced(cur_lane.dist_hub, SpeedMode.HUB, assist_active = AssistActive.DYNAMIC)
-                self.afcDeltaTime.log_with_time("Loaded to hub")
+                dist_to_hub = cur_lane.dist_hub
+                if (self.homing_enabled
+                    and cur_lane.is_direct_hub()):
+                    dist_to_hub += cur_hub.move_dis
+
+                home_endstop = AFCHomingPoints.HUB
+                if cur_lane.is_direct_hub():
+                    home_endstop= cur_lane.get_toolhead_endstop()
+
+                cur_lane.move_to(dist_to_hub, SpeedMode.HUB,
+                                 assist_active=AssistActive.DYNAMIC,
+                                 endstop=home_endstop, use_homing=self.homing_enabled)
+                self.afcDeltaTime.log_with_time(
+                    f"Loaded to {'hub' if cur_lane.hub != 'direct' else 'toolhead'}"
+                )
 
             cur_lane.loaded_to_hub = True
             hub_attempts = 0
 
+            if cur_lane.hub != 'direct' and not cur_hub.state:
+                cur_lane.move_to(cur_hub.move_dis, SpeedMode.SHORT,
+                                 endstop=AFCHomingPoints.HUB,
+                                 use_homing=self.homing_enabled)
+                hub_attempts += 1
+
             # Ensure filament moves past the hub.
             while not cur_hub.state and not cur_lane.is_direct_hub():
-                if hub_attempts == 0:
-                    cur_lane.move_advanced(cur_hub.move_dis, SpeedMode.SHORT)
-                else:
-                    cur_lane.move_advanced(cur_lane.short_move_dis, SpeedMode.SHORT)
+                cur_lane.move_to(cur_hub.move_dis, SpeedMode.SHORT,
+                                 endstop=AFCHomingPoints.HUB,
+                                 use_homing=self.homing_enabled and self.home_to_hub)
                 hub_attempts += 1
                 if hub_attempts > 20:
                     message = 'filament did not trigger hub sensor, CHECK FILAMENT PATH\n||=====||==>--||-----||\nTRG   LOAD   HUB   TOOL.'
@@ -1343,15 +1383,27 @@ class afc:
 
             # Move filament towards the toolhead.
             if not cur_lane.is_direct_hub():
-                cur_lane.move_advanced(cur_hub.afc_bowden_length, SpeedMode.LONG, assist_active = AssistActive.YES)
+                cur_lane.move_to(cur_hub.afc_bowden_length,
+                                 SpeedMode.LONG,
+                                 assist_active=AssistActive.YES,
+                                 endstop=cur_lane.get_toolhead_endstop(),
+                                 use_homing=self.homing_enabled and self.home_to_tool)
 
             # Ensure filament reaches the toolhead.
             tool_attempts = 0
             if cur_extruder.tool_start:
                 while not cur_lane.get_toolhead_pre_sensor_state():
                     tool_attempts += 1
-                    cur_lane.move(cur_lane.short_move_dis, cur_extruder.tool_load_speed, cur_lane.long_moves_accel)
-                    if tool_attempts > int(self.tool_homing_distance/cur_lane.short_move_dis):
+                    move_distance = cur_lane.short_move_dis
+                    max_attempts = int(self.tool_homing_distance/cur_lane.short_move_dis)
+                    if (self.homing_enabled
+                        and self.home_to_tool):
+                        move_distance = cur_hub.afc_bowden_length
+                        max_attempts = 2
+                    cur_lane.move_to(move_distance, SpeedMode.SHORT,
+                                     endstop=cur_lane.get_toolhead_endstop(),
+                                     use_homing=self.homing_enabled and self.home_to_tool)
+                    if tool_attempts >= max_attempts:
                         message = 'filament failed to trigger pre extruder gear toolhead sensor, CHECK FILAMENT PATH\n||=====||====||==>--||\nTRG   LOAD   HUB   TOOL'
                         message += '\nTo resolve set lane loaded with `SET_LANE_LOADED LANE={}` macro.'.format(cur_lane.name)
                         message += '\nManually move filament with LANE_MOVE macro for {} until filament is right before toolhead extruder gears,'.format(cur_lane.name)
@@ -1371,7 +1423,8 @@ class afc:
             if cur_extruder.tool_end:
                 while not cur_extruder.tool_end_state:
                     tool_attempts += 1
-                    self.move_e_pos( cur_lane.short_move_dis, cur_extruder.tool_load_speed, "Tool end", wait_tool=True )
+                    self.move_e_pos(cur_lane.short_move_dis, cur_extruder.tool_load_speed,
+                                    "Tool end", wait_tool=True )
                     if tool_attempts > 20:
                         message = 'filament failed to trigger post extruder gear toolhead sensor, CHECK FILAMENT PATH\n||=====||====||==>--||\nTRG   LOAD   HUB   TOOL'
                         message += '\nTo resolve set lane loaded with `SET_LANE_LOADED LANE={}` macro.'.format(cur_lane.name)
@@ -1413,6 +1466,7 @@ class afc:
 
         # Update tool and lane status.
         cur_lane.set_tool_loaded()
+        cur_lane.espooler.do_assist_move()
         # Setting disable_fault so that fault detection is turned off for users
         # that utilize poop
         cur_lane.enable_buffer(disable_fault=True)
@@ -1717,9 +1771,15 @@ class afc:
             # Synchronize and move filament out of the hub.
             cur_lane.unsync_to_extruder()
             if not cur_lane.is_direct_hub():
-                cur_lane.move_advanced(cur_hub.afc_unload_bowden_length * -1, SpeedMode.LONG, assist_active = AssistActive.YES)
+                cur_lane.move_to(cur_hub.afc_unload_bowden_length * -1, SpeedMode.LONG,
+                                assist_active=AssistActive.YES,
+                                endstop=AFCHomingPoints.HUB,
+                                use_homing=self.homing_enabled)
             else:
-                cur_lane.move_advanced(cur_lane.dist_hub * -1, SpeedMode.HUB, assist_active = AssistActive.DYNAMIC)
+                cur_lane.move_to(cur_lane.dist_hub * -1, SpeedMode.LONG,
+                                 assist_active = AssistActive.DYNAMIC,
+                                 endstop=AFCHomingPoints.LOAD,
+                                 use_homing=self.homing_enabled)
 
             self.afcDeltaTime.log_with_time("Long retract done")
 
@@ -1730,9 +1790,17 @@ class afc:
             # Ensure filament is fully cleared from the hub.
             num_tries = 0
             while cur_hub.state:
-                cur_lane.move_advanced(cur_lane.short_move_dis * -1, SpeedMode.SHORT, assist_active = AssistActive.YES)
+                max_attempts = (cur_hub.afc_unload_bowden_length / cur_lane.short_move_dis)
+                move_dist = cur_lane.short_move_dis
+                if self.homing_enabled:
+                    max_attempts = 2
+                    move_dist = cur_hub.afc_unload_bowden_length
+                cur_lane.move_to(move_dist * -1, SpeedMode.SHORT,
+                                 assist_active=AssistActive.YES,
+                                 endstop=AFCHomingPoints.HUB,
+                                 use_homing=self.homing_enabled)
                 num_tries += 1
-                if num_tries > (cur_hub.afc_unload_bowden_length / cur_lane.short_move_dis):
+                if num_tries >= max_attempts:
                     # Handle failure if the filament doesn't clear the hub.
                     message = 'Hub is not clearing, filament may be stuck in hub'
                     message += '\nPlease check to make sure filament has not broken off and caused the sensor to stay stuck'
@@ -1751,7 +1819,8 @@ class afc:
 
             #Move to make sure hub path is clear based on the move_clear_dis var
             if not cur_lane.is_direct_hub():
-                cur_lane.move_advanced(cur_hub.hub_clear_move_dis * -1, SpeedMode.SHORT, assist_active = AssistActive.YES)
+                cur_lane.move_advanced(cur_hub.hub_clear_move_dis * -1, SpeedMode.SHORT,
+                                       assist_active = AssistActive.YES)
 
                 # Cut filament at the hub, if configured.
                 if cur_hub.cut:
@@ -1762,9 +1831,9 @@ class afc:
 
                     # Confirm the hub is clear after the cut.
                     while cur_hub.state:
-                        cur_lane.move_advanced(cur_lane.short_move_dis * -1, SpeedMode.SHORT, assist_active = AssistActive.YES)
+                        cur_lane.move_advanced(cur_lane.short_move_dis * -1, SpeedMode.SHORT,
+                                               assist_active = AssistActive.YES)
                         num_tries += 1
-                        # TODO: Figure out max number of tries
                         if num_tries > (cur_hub.afc_unload_bowden_length / cur_lane.short_move_dis):
                             message = 'HUB NOT CLEARING after hub cut\n'
                             self.error.handle_lane_failure(cur_lane, message)
@@ -1779,7 +1848,8 @@ class afc:
 
             if cur_lane.is_direct_hub():
                 while cur_lane.load_state:
-                    cur_lane.move_advanced(cur_lane.short_move_dis * -1, SpeedMode.SHORT, assist_active = AssistActive.YES)
+                    cur_lane.move_advanced(cur_lane.short_move_dis * -1, SpeedMode.SHORT,
+                                           assist_active = AssistActive.YES)
                 cur_lane.move_advanced(cur_lane.short_move_dis * -5, SpeedMode.SHORT)
 
             if self.post_unload_macro is not None: