+++ extras/AFC_functions.py
@@ -20,19 +20,24 @@ from configfile import error
 from datetime import datetime
 from pathlib import Path
 
-from typing import TYPE_CHECKING, Optional
+from typing import TYPE_CHECKING, Optional, Union
 
 if TYPE_CHECKING:
     from extras.AFC import afc
     from AFC_logger import AFC_logger
     from extras.AFC_lane import AFCLane
     from extras.AFC_extruder import AFCExtruder
+    from extras.AFC_stepper import AFCExtruderStepper
+    from extras.AFC_hub import afc_hub
 
 try: from extras.AFC_utils import ERROR_STR
 except: raise error("Error when trying to import AFC_utils.ERROR_STR\n{trace}".format(trace=traceback.format_exc()))
 
 try: from extras.AFC_respond import AFCprompt
 except: raise error(ERROR_STR.format(import_lib="AFC_respond", trace=traceback.format_exc()))
+
+try: from extras.AFC_lane import SpeedMode, AssistActive, AFCHomingPoints
+except: raise error(ERROR_STR.format(import_lib="AFC_lane", trace=traceback.format_exc()))
 
 def load_config(config):
     return afcFunction(config)
@@ -1396,7 +1401,7 @@ class afcFunction:
                                   "LANE": {"default": "lane1", "type": "string"}}
     def cmd_AFC_LANE_RESET(self, gcmd):
         """
-        This function resets a specified lane to the hub position in the AFC system. It checks for various error conditions,
+        This macro resets a specified lane to the hub position in the AFC system. It checks for various error conditions,
         such as whether the toolhead is loaded or whether the hub is already clear. The function moves the lane back to the
         hub based on the specified or default distances, ensuring the lane's correct state before completing the reset.
 
@@ -1418,7 +1423,7 @@ class afcFunction:
 
         prompt = AFCprompt(gcmd, self.logger)
         lane = gcmd.get('LANE', None)
-        long_dis = gcmd.get('DISTANCE', None)
+        long_dis = gcmd.get('DISTANCE', 50)
 
         if not lane:
             prompt.p_end()
@@ -1440,8 +1445,8 @@ class afcFunction:
                 self.afc.error.AFC_error("DISTANCE must be a valid number.", pause=False)
                 return
 
-        cur_lane = self.afc.lanes[lane]
-        CUR_HUB = cur_lane.hub_obj
+        cur_lane: Union[AFCLane, AFCExtruderStepper] = self.afc.lanes[lane]
+        CUR_HUB: afc_hub = cur_lane.hub_obj
         short_move = cur_lane.short_move_dis * 2
 
         if not CUR_HUB.state:
@@ -1459,7 +1464,11 @@ class afcFunction:
         fail_state_msg = "'{}' failed to reset to hub, {} switch became false during reset"
 
         if long_dis is not None:
-            cur_lane.move(float(long_dis) * -1, cur_lane.long_moves_speed, cur_lane.long_moves_accel, True)
+            cur_lane.move_to(distance=float(long_dis) *-1,
+                             speed_mode=SpeedMode.SHORT,
+                             endstop=AFCHomingPoints.HUB,
+                             assist_active=AssistActive.YES,
+                             use_homing=self.afc.homing_enabled)
 
         while CUR_HUB.state:
             cur_lane.move(short_move * -1, cur_lane.short_moves_speed, cur_lane.short_moves_accel, True)
@@ -1477,7 +1486,7 @@ class afcFunction:
                 self.afc.error.AFC_error("'{}' failed to reset to hub".format(cur_lane), pause=False)
                 return
 
-        cur_lane.move(CUR_HUB.move_dis * -1, cur_lane.short_moves_speed, cur_lane.short_moves_accel, True)
+        cur_lane.move(CUR_HUB.hub_clear_move_dis * -1, cur_lane.short_moves_speed, cur_lane.short_moves_accel, True)
         cur_lane.loaded_to_hub = True
         cur_lane.do_enable(False)